#5270. Maintain number of free events available in events pool.
Unfortunately, idLinkList does not maintain the number of elements in it.
Computing it on every event allocation would mean that growth to N events would take O(N^2) time, which is not nice.
So I try to maintain it as separate global variable.

There is an assert which checks it validity, and I'm not entirely sure it will never fail.
For some reason EventQueue.size() + FreeEvents.size() is not always MAX_EVENTS, sometimes it is one event off.
To avoid slowing down debug build, validity check works only when the number of alive events is small (i.e. in normal condition).