#5170. Fix for crash with "decl_stack 1" on loading of "The Rat's Triumphant". The reason was dangling pointers to idWindow.
Unlike all the other objects, idWindow-s seem to be recreated and deleted too eagerly.
For now I have "fixed" it with a very dirty hack.
When LoadStack needs to print window level, it asks global uiManager if this window pointer is currently present in all of its GUIs.
If it is not, then name is not queried.

It means that sometimes the name of the window will be missing in the load stack.
There is also small probability that window name will be reported incorrectly, if the same memory buffer will be reused for another window.
But, most importantly, it does not crash now.