#5138. Changed the way how randomizer is computed for particle deform systems.
Changed the way how randomizer is computed (see also #4132).
It no longer depends on bounding box coordinates.
Instead, it depends on "signature" (idPartSysSurfaceEmitterSignature) of the system, which includes model name, surface index, and particle stage index.

The signature is expected to be unique identifier of a system (must be so for particle deform on brushes/patches).
It is also used to deduce cutoff image name for collisionStatic feature.

Added the new function to renderer/ParticleSystem.[h|cpp]:
  float idParticle_ComputeSurfaceRandomizer(const idPartSysSurfaceEmitterSignature &sign, float diversity);
"diversity" is taken from entity definition (if present).