#5437. Implemented "mapLayout linear" for particle collisions.
Note that the "mapLayout linear" works under conditions:
  "diversityPeriod" is set (size of precomputed texture is proportional to this value)
  "collisionStaticTimeSteps" is set, unless every particle moves along perfect line (precomputation time is proportional to this value)

The "collisionStaticTimeSteps" parameter specifies how many segments would be in the polyline approximating particle trajectory in runParticle.
This number can be omitted only for particles which move perfectly linearly, for which one ray trace is enough to find collision.
If particle moves along curved path, then its path is approximated with polyline, which is intersected with the world.

-----------------

Had to expose the following function in renderer/ParticleSystem_[decl/def].h
  int idParticle_GetRandomSeed(PIN(int) index, PIN(int) cycleNumber, PIN(float) randomizer);
It allows to compute random seed in runParticle tool, without worrying about various age cutoffs in idParticle_EmitParticle.

Main changes in tools/compilers/particle/ParticleCollisionStatic.cpp:
  Extracted actual particle modeling/collision code into PrtCollision::ModelParticleMovement
  Implemented code for "mapLayout linear" in PrtCollision::ProcessSurfaceEmitter
  Pass "shaderParm5" spawnarg as diversity for entities (but not the world)

The runtime usage of linear cutoff map is added to renderer/tr_deform.cpp.
The core functions are in idParticle_FetchCutoffTimeLinear, and partly in idParticle_PrepareCutoffMap.