#5542. Now script-generated frob never triggers Use-By-Frobbing behavior.
The main frobbing method idPlayer::PerformFrob now accepts a bool, which says whether use-by-frobbing is allowed or not.
If it is allowed and enabled by player in gameplay menu, then the behavior triggers.
Note that idEntity::Event_Frob passes false, so it never leads to "use current inventory item on frobbed object".