#5551. Force curl to return error when remote file is not available (and on other HTTP errors).
By default HTTP error (code >= 400) is not considered an error of curl itself, it just returns CURLE_OK.
After that SHA verification of FM.pk4 kicks in and complains to user that mission information is wrong.

Now if file is not present on mirror (not synchronized yet), then curl returns CURLE_HTTP_RETURNED_ERROR.
Then SHA is not verified, and thus downloader tries the next URL in the list.