#5683. Support changing LOD settings in hot-reload.
If any of the hot-reload spawnargs are modified, then we call idGameEdit::EntityUpdateLOD.
This function:
  1) deletes old LOD data
  2) creates new LOD data by parsing new spawnargs from scratch (with ParseLODSpawnargs)
  3) forces updating LOD state right now (with SwitchLOD)
Some changes were necessary to ensure that SwitchLOD works with hot-reload.

First of all, it must work normally if LOD data is missing (after user has deleted LOD spawnargs from entity).
It means returning back the original 0-th level of LOD in model, skin, noshadows, offset.
Note that restoring original skin in this case is not always correct (ignores "random_skin" and default skin of modelDef).

One major issues was with offset_lod_X spawnarg.
First of all, there was a minor bug in checking whether we need to call SetOrigin or not (basically, it was almost always called).
Also, now we store current offset in entity members, since otherwise it is impossible to restore proper offset if user has changed them in map and hot-reloaded the entity.