#5683. Supported hot-reloading moved/rotated entities bound to master in a simple way.
Unfortunately, when entity is bound to something, SetOrigin method accepts data in local coordinate system of the master, but .map data is in global coords.
So we detect master entity, read its new "origin"/"rotation"/"angle", and convert data of the bound entity into local coords.
Supports even moving a bunch of entities bound to each other --- all at once.

Note that we only support this in a simple case, when:
1) bindToBody and bindToJoint are missing (i.e. not attached to skeleton or multi-clipmodel).
2) The entity is still bound to the same entity as in the .map file, and that entity still exists.
If any condition is not satisfied, then we unbind the entity and move/rotate normally.