#5816. Volumetric lights are now randomly sampled, number of samples can no longer be set in material.
First of all, two cvars are added:
  r_volumetricSamples --- number of samples used
  r_volumetricRandomize --- whether to apply randomization (see below)
Consequently, volumetricLight no longer accepts parameter.

Randomization does the following.
Imagine that we have a segment within light polytope which needs to be sampled, and we split it into N subsegments.
Without randomization, we take a sample in the middle of each subsegment.
With randomization, random sample is chosed on each subsegment.
Random value is dependent on screen position and sample index.