Fixed bug caused by svn rev 9698: don't try to set stencil textureing mode if stencil texture is not generated yet.
Such situation can easily happen when shadow maps are enabled, if all lights on the map use them.
In such case no light uses stencil shadows, so they are not generated, so shadow FBO is not generated, so the stencil texture is not generated.
Trying to set stencil texturing mode causes OpenGL warnings spam (unless they are silenced, of course).